+++ packages/shared-ux/error_boundary/src/ui/error_messages.tsx
@@ -8,6 +8,8 @@
 
 import React from 'react';
 
+import { FormattedMessage } from '@kbn/i18n-react';
+
 import {
   EuiAccordion,
   EuiButton,
@@ -15,10 +17,15 @@ import {
   EuiCode,
   EuiCodeBlock,
   EuiEmptyPrompt,
+  EuiFlexGroup,
+  EuiFlexItem,
   EuiPanel,
   EuiSpacer,
   useGeneratedHtmlId,
 } from '@elastic/eui';
+
+const DATA_TEST_SUBJ_PAGE_REFRESH_BUTTON = 'pageReloadButton';
+const DATA_TEST_SUBJ_PAGE_DETAILS_BUTTON = 'showDetailsButton';
 
 export interface ErrorCalloutProps {
   error: Error;
@@ -27,7 +34,7 @@ export interface ErrorCalloutProps {
   reloadWindow: () => void;
 }
 
-export const ErrorCallout = (props: ErrorCalloutProps) => {
+export const FatalPrompt = (props: ErrorCalloutProps) => {
   const { error, errorInfo, name: errorComponentName, reloadWindow } = props;
   const errorBoundaryAccordionId = useGeneratedHtmlId({ prefix: 'errorBoundaryAccordion' });
   return (
@@ -56,7 +63,60 @@ export const ErrorCallout = (props: ErrorCalloutProps) => {
   );
 };
 
-export const RefresherPrompt = (props: ErrorCalloutProps) => {
+export const FatalInline = (_props: ErrorCalloutProps) => {
+  return <EuiCallOut color="danger" iconType="error" title="Error: unable to load." />;
+};
+
+export const FatalToastTitle = () => (
+  <FormattedMessage
+    id="sharedUXPackages.error_boundary.fatal.toastError.title"
+    defaultMessage="A fatal error was encountered."
+  />
+);
+
+export const FatalToastText = ({ reloadWindow }: { reloadWindow: () => void }) => {
+  return (
+    <EuiFlexGroup direction="column">
+      <EuiFlexItem>
+        <p>Try refreshing the page.</p>
+      </EuiFlexItem>
+      <EuiFlexItem>
+        <EuiFlexGroup gutterSize="s" direction="row">
+          <EuiFlexItem>
+            <EuiButton
+              size="s"
+              onClick={() => {}}
+              data-test-subj={DATA_TEST_SUBJ_PAGE_DETAILS_BUTTON}
+              fill={false}
+              color="danger"
+            >
+              <FormattedMessage
+                id="sharedUXPackages.error_boundary.toastError.fatal.details"
+                defaultMessage="Show details"
+              />
+            </EuiButton>
+          </EuiFlexItem>
+          <EuiFlexItem>
+            <EuiButton
+              size="s"
+              onClick={reloadWindow}
+              data-test-subj={DATA_TEST_SUBJ_PAGE_REFRESH_BUTTON}
+              fill={true}
+              color="danger"
+            >
+              <FormattedMessage
+                id="sharedUXPackages.error_boundary.toastError.fatal.pageRefreshButtonLabel"
+                defaultMessage="Refresh"
+              />
+            </EuiButton>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      </EuiFlexItem>
+    </EuiFlexGroup>
+  );
+};
+
+export const RecoverablePrompt = (props: ErrorCalloutProps) => {
   const { reloadWindow } = props;
   return (
     <EuiEmptyPrompt
@@ -70,5 +130,43 @@ export const RefresherPrompt = (props: ErrorCalloutProps) => {
         </EuiButton>
       }
     />
+  );
+};
+
+export const RecoverableInline = (props: ErrorCalloutProps) => {
+  return <EuiCallOut iconType="broom" title="Please refresh." onClick={props.reloadWindow} />;
+};
+
+export const RecoverableToastTitle = () => (
+  <FormattedMessage
+    id="sharedUXPackages.error_boundary.toastError.recoverable.title"
+    defaultMessage="Sorry, please refresh"
+  />
+);
+
+export const RecoverableToastText = ({ reloadWindow }: { reloadWindow: () => void }) => {
+  return (
+    <EuiFlexGroup direction="column">
+      <EuiFlexItem>
+        <p>An error occurred when trying to load a part of the page. Please try refreshing.</p>
+      </EuiFlexItem>
+      <EuiFlexItem>
+        <EuiFlexGroup gutterSize="s" direction="row">
+          <EuiFlexItem>
+            <EuiButton
+              size="s"
+              onClick={reloadWindow}
+              data-test-subj={DATA_TEST_SUBJ_PAGE_REFRESH_BUTTON}
+              fill={true}
+            >
+              <FormattedMessage
+                id="sharedUXPackages.error_boundary.toastError.recoverable.pageRefreshButtonLabel"
+                defaultMessage="Refresh"
+              />
+            </EuiButton>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      </EuiFlexItem>
+    </EuiFlexGroup>
   );
 };
